scraper/requirements.txt
beautifulsoup4>=4.12, <4.13
requests>=2.31, <2.32
pypdf2>=3.0, <3.1
openai>=0.27, <0.28
tiktoken>=0.4, <0.5
python-dotenv>= 1.0, <1.1
redis-om>=0.1, <0.2
